Moved code to activate at different times
Code clean-ups etc...
