+++ changelog.txt
@@ -1,4 +1,11 @@
 ---------------------------------------------------------------------------------------------------
+Version: 0.18.6
+Date: 2020-04-03
+  Bugfixes:
+    - Changed load ordering to make the mod more compatible.
+    - Code clean-up and moved functions out of staging files.
+
+---------------------------------------------------------------------------------------------------
 Version: 0.18.5
 Date: 2020-03-21
   Bugfixes:
