+++ data.lua
@@ -1,0 +1,2 @@
+require("prototypes.LogisticsUpdates")
+run_early_updates()
